change poststatus.id to SMALLINT to match post.status
git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@1570 653ae4dd-d31e-0410-96ef-6bf7bf53c507

+++ schema/mysql/schema.sql
@@ -30,7 +30,7 @@ CREATE TABLE  {$prefix}posttype (
 );
 
 CREATE TABLE  {$prefix}poststatus (
-  id INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  id SMALLINT UNSIGNED NOT NULL AUTO_INCREMENT,
   name VARCHAR(255) NOT NULL,
   internal TINYINT(1),
   PRIMARY KEY (id)
